+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 Changes:
 
+* Added single Edit Dilution page
 * Changed Edit Library page to show validation messages instead of error page when saving fails
 * Fixed missing Notes section on Edit Sample page